Move theme into separate component
