+++ frontend/.dockerignore
@@ -5,3 +5,4 @@ npm-debug.log
 README.md
 .next
 .git
+**/*.swp
