+++ frontend/pages/_app.tsx
@@ -2,7 +2,7 @@ import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 import { ChakraProvider } from '@chakra-ui/react'
 import '@fontsource/inter'
-import theme from './theme'
+import theme from '../components/theme'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
