+++ frontend/pages/_document.tsx
@@ -1,6 +1,6 @@
 import { Html, Head, Main, NextScript } from 'next/document'
 import { ColorModeScript } from '@chakra-ui/react'
-import theme from './theme'
+import theme from '../components/theme'
 
 export default function Document() {
   return (
